backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Anthropic Claude API
anthropic>=0.53.0

# Async support
httpx==0.26.0
aiofiles==23.2.1

# Logging and monitoring
structlog==24.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11

# pyRevit communication
websockets==12.0